chore(generate): move command under `cli-commands` folder
cause wuf is not only cli
