+++ packages/wix-ui-framework/src/cli-commands/generate/__test__/copy-templates.spec.ts
