+++ packages/wix-ui-framework/src/cli-commands/generate/__test__/replace-templates.spec.ts
