+++ packages/wix-ui-framework/src/cli-commands/generate/__test__/verify-working-directory.spec.ts
