+++ packages/wix-ui-framework/src/cli-commands/generate/__testfixtures__/index.output.js
