+++ packages/wix-ui-framework/src/cli-commands/generate/__testfixtures__/testkit-definitions.input.js
