+++ packages/wix-ui-framework/src/cli-commands/generate/__testfixtures__/testkit-protractor.output.js
