+++ packages/wix-ui-framework/src/cli-commands/generate/__testfixtures__/testkit-puppeteer.input.js
