--- packages/wix-ui-framework/src/generate/__test__/templates/test-generated/src/Component/Component.meta.js
@@ -1,8 +1,0 @@
-import {%ComponentName%} from './{%ComponentName%}';
-import Registry from '@ui-autotools/registry';
-
-const metadata = Registry.getComponentMetadata({%ComponentName%});
-
-metadata.exportedFrom({
-  path: 'src/{%ComponentName%}/{%ComponentName%}.js',
-});
