--- packages/wix-ui-framework/src/generate/__test__/templates/test-generated/src/Component/Component.st.css
@@ -1,25 +1,0 @@
-:import {
-  -st-from: "wix-ui-core/dist/src/themes/backoffice/colors.st.css";
-  -st-named: G30,R30;
-}
-
-.root {
-  text-align: center;
-}
-
-.number {
-  -st-states: odd, even;
-}
-
-.number:even {
-  background: value(G30);
-}
-
-.number:odd {
-  background: value(R30);
-}
-
-.button {
-  display: flex;
-  justify-content: space-evenly;
-}
