--- packages/wix-ui-framework/src/generate/__test__/templates/test-generated/src/Component/docs/index.story.js
@@ -1,97 +1,0 @@
-import React from 'react';
-import {
-  header,
-  tabs,
-  tab,
-  description,
-  importExample,
-  title,
-  columns,
-  divider,
-  code as baseCode,
-  playground,
-  api,
-  testkit,
-} from 'wix-storybook-utils/Sections';
-
-import { storySettings } from '../test/storySettings';
-import allComponents from '../../../stories/utils/allComponents';
-
-import {%ComponentName%} from '..';
-
-const code = config => baseCode({ components: allComponents, ...config });
-
-export default {
-  category: storySettings.category,
-  storyName: '{%ComponentName%}',
-
-  component: {%ComponentName%},
-  componentPath: '..',
-
-  componentProps: {
-    buttonText: 'Hello World!',
-  },
-
-  exampleProps: {
-    // Put here presets of props, for more info:
-    // https://github.com/wix/wix-ui/blob/master/packages/wix-storybook-utils/docs/usage.md#using-list
-  },
-
-  sections: [
-    header({
-      issueUrl: 'https://github.com/wix/wix-style-react/issues/new',
-      sourceUrl:
-        'https://github.com/wix/wix-style-react/tree/master/src/{%ComponentName%}/',
-      component: <{%ComponentName%} buttonText="Click me!" />,
-    }),
-
-    tabs([
-      tab({
-        title: 'Description',
-        sections: [
-          columns([
-            description({
-              title: 'Description',
-              text:
-                'This line here should briefly describe component in just a sentence or two. It should be short and easy to read.',
-            }),
-          ]),
-
-          columns([
-            importExample(
-              "import {%ComponentName%} from 'wix-style-react/{%ComponentName%}';",
-            ),
-          ]),
-
-          divider(),
-
-          title('Examples'),
-
-          columns([
-            description({
-              title: 'Simple Usage',
-              text: 'A simple example with compact preview',
-            }),
-
-            code({
-              compact: true,
-              source: '<{%ComponentName%} buttonText="Hello World!"/>',
-            }),
-          ]),
-
-          code({
-            title: 'Full Interactive Preview',
-            description: 'A non compact version of same code example as above',
-            source: '<{%ComponentName%} buttonText="Hello World!"/>',
-          }),
-        ],
-      }),
-
-      ...[
-        { title: 'API', sections: [api()] },
-        { title: 'Testkit', sections: [testkit()] },
-        { title: 'Playground', sections: [playground()] },
-      ].map(tab),
-    ]),
-  ],
-};
