--- packages/wix-ui-framework/src/generate/__test__/templates/test-generated/src/Component/index.js
@@ -1,1 +1,0 @@
-export { default } from './{%ComponentName%}.js';
