--- packages/wix-ui-framework/src/generate/__test__/templates/test-generated/src/Component/test/Component.private.uni.driver.js
@@ -1,9 +1,0 @@
-import { {%componentName%}DriverFactory as publicDriverFactory } from '../{%ComponentName%}.uni.driver';
-
-export const {%componentName%}PrivateDriverFactory = base => {
-  return {
-    ...publicDriverFactory(base),
-
-    // Add here driver methods that considered "private"
-  };
-};
