--- packages/wix-ui-framework/src/generate/__test__/templates/test-generated/src/Component/test/ComponentStories.js
@@ -1,16 +1,0 @@
-import React from 'react';
-import { storiesOf } from '@storybook/react';
-import { getTestStoryKind } from '../../../stories/storiesHierarchy';
-import { storySettings } from './storySettings';
-
-import {%ComponentName%} from '..';
-
-const TestStories = storiesOf(getTestStoryKind(storySettings), module);
-const { testStoryNames, dataHook } = storySettings;
-
-TestStories.add(testStoryNames.DEFAULT, () => (
-  <{%ComponentName%}
-    dataHook={dataHook}
-    buttonText="Press me for a surprise"
-  />
-));
