--- packages/wix-ui-framework/src/generate/__test__/templates/test-generated/src/Component/test/storySettings.js
@@ -1,10 +1,0 @@
-import { Category } from '../../../stories/storiesHierarchy';
-
-export const storySettings = {
-  category: Category.COMPONENTS,
-  storyName: '{%ComponentName%}',
-  dataHook: 'story-{%component-name%}',
-  testStoryNames: {
-    DEFAULT: 'Default',
-  },
-};
