--- packages/wix-ui-framework/src/generate/__test__/templates/test-generated/src/index.js
@@ -1,1 +1,0 @@
-// explicit test file
