--- packages/wix-ui-framework/src/generate/__test__/templates/test-generated/testkit/protractor.js
@@ -1,1 +1,0 @@
-// explicit test file
