+++ packages/wix-ui-framework/tsconfig.json
@@ -25,8 +25,8 @@
     "./src/**/*.js"
   ],
   "exclude": [
-    "src/generate/__test__/templates/**/*.*",
-    "src/generate/__testfixtures__/**/*.*",
+    "src/cli-commands/generate/__test__/templates/**/*.*",
+    "src/cli-commands/generate/__testfixtures__/**/*.*",
     "./package.json"
   ]
 }
